requirements/base.txt
# asgiref==3.4.1
# autopep8==1.5.7
# certifi==2021.10.8
# cffi==1.15.0
# cryptography==36.0.1
# defusedxml==0.7.1
# dj-database-url==0.5.0
# Django==3.2.6
# django-allauth==0.47.0
# django-ckeditor==6.2.0
django-compressor
django-crispy-forms==1.13.0
# django-extensions==3.1.5
# django-js-asset==1.2.2
# django-taggit==2.0.0
# djangorestframework==3.13.1
# idna==3.3
# jsmin==3.0.0
# oauthlib==3.1.1
Pillow==8.4.0
# python-decouple==3.4
# pytz==2021.1
# rcssmin==1.1.0
# rjsmin==1.2.0
# six==1.16.0
# social-auth-core==4.1.0
# Unipath==1.1

django-extensions==3.2.1
python-decouple==3.8
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.2